STEWARDS_multi_daily_post2009.R
R code for processing Micronet data.  Grabs the date from the file names of all text files in the directory, creates a column of these values, then concatenates all the files into one large table with the "date" filename column.  It then puts this into a more usable format and creates a .csv file for adding to STEWARDS databases.